HierCDF local test passed (train, eval)
